backend/requirements.txt
asgiref==3.9.1
Django==5.2.4
django-cors-headers==4.7.0
django-environ==0.12.0
djangorestframework==3.16.0
djangorestframework_simplejwt==5.5.0
PyJWT==2.9.0
sqlparse==0.5.3
tzdata==2025.2
uvicorn==0.34.3
channels==4.2.0
channels-redis==4.2.1
django-redis==6.0.0
psycopg2-binary==2.9.10
python-decouple==3.8